requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
django-extensions==3.2.3 ; python_version >= "3.11" and python_version < "4.0"
django-filter==23.5 ; python_version >= "3.11" and python_version < "4.0"
django==5.0.3 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework-jsonapi==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
